+++ apps/www/src/registry/components/editor/use-create-editor.ts
@@ -14,6 +14,7 @@ import {
   UnderlinePlugin,
 } from '@udecode/plate-basic-marks/react';
 import { BlockquotePlugin } from '@udecode/plate-block-quote/react';
+import { CalloutPlugin } from '@udecode/plate-callout/react';
 import {
   CodeBlockPlugin,
   CodeLinePlugin,
@@ -64,6 +65,7 @@ import {
 import { AIAnchorElement } from '@/registry/ui/ai-anchor-element';
 import { AILeaf } from '@/registry/ui/ai-leaf';
 import { BlockquoteElement } from '@/registry/ui/blockquote-element';
+import { CalloutElement } from '@/registry/ui/callout-element';
 import { CodeBlockElement } from '@/registry/ui/code-block-element';
 import { CodeLeaf } from '@/registry/ui/code-leaf';
 import { CodeLineElement } from '@/registry/ui/code-line-element';
@@ -108,6 +110,7 @@ export const viewComponents = {
   [AudioPlugin.key]: MediaAudioElement,
   [BlockquotePlugin.key]: BlockquoteElement,
   [BoldPlugin.key]: withProps(PlateLeaf, { as: 'strong' }),
+  [CalloutPlugin.key]: CalloutElement,
   [CodeBlockPlugin.key]: CodeBlockElement,
   [CodeLinePlugin.key]: CodeLineElement,
   [CodePlugin.key]: CodeLeaf,