+++ packages/markdown/src/lib/rules/types.ts
@@ -10,6 +10,7 @@ import type {
   TEquationElement,
   TImageElement,
   TLinkElement,
+  TMediaElement,
   TMentionElement,
   TSuggestionText,
   TTableCellElement,
@@ -47,11 +48,6 @@ import type {
 import type { SerializeMdOptions } from '../serializer';
 
 /* eslint-disable perfectionist/sort-object-types */
-export type TRules = Partial<{
-  [K in keyof PlateNodeTypeMap]: Nullable<TNodeParser<K>>;
-}> &
-  Record<string, Nullable<AnyNodeParser>>;
-
 export type AnyNodeParser = {
   mark?: boolean;
   deserialize?: (
@@ -74,6 +70,11 @@ export type TNodeParser<K extends keyof PlateNodeTypeMap> = {
     options: SerializeMdOptions
   ) => MdastNodeTypeMap[K];
 };
+
+export type TRules = Partial<{
+  [K in keyof PlateNodeTypeMap]: Nullable<TNodeParser<K>>;
+}> &
+  Record<string, Nullable<AnyNodeParser>>;
 
 type MdastNodeTypeMap = {
   /** Common Elements */
@@ -126,6 +127,9 @@ type MdastNodeTypeMap = {
   html: MdHtml;
 
   /** PlateOnly */
+  file: any;
+  video: any;
+  audio: any;
 };
 
 type PlateNodeTypeMap = {
@@ -179,5 +183,9 @@ type PlateNodeTypeMap = {
   imageReference: any;
   linkReference: any;
   html: any;
-  /** PlateOnly */
+
+  /** PlateOnly media */
+  file: TMediaElement;
+  video: TMediaElement;
+  audio: TMediaElement;
 };