+++ templates/plate-playground-template/src/components/plate-ui/media-audio-element.tsx
@@ -11,31 +11,29 @@ import { Caption, CaptionTextarea } from './caption';
 
 export const MediaAudioElement = withHOC(
   ResizableProvider,
-  withRef<typeof PlateElement>(
-    ({ children, className, nodeProps, ...props }, ref) => {
-      const { align = 'center', readOnly, unsafeUrl } = useMediaState();
+  withRef<typeof PlateElement>(({ children, className, ...props }, ref) => {
+    const { align = 'center', readOnly, unsafeUrl } = useMediaState();
 
-      return (
-        <PlateElement ref={ref} className={cn(className, 'mb-1')} {...props}>
-          <figure
-            className="group relative cursor-default"
-            contentEditable={false}
-          >
-            <div className="h-16">
-              <audio className="size-full" src={unsafeUrl} controls />
-            </div>
+    return (
+      <PlateElement ref={ref} className={cn(className, 'mb-1')} {...props}>
+        <figure
+          className="group relative cursor-default"
+          contentEditable={false}
+        >
+          <div className="h-16">
+            <audio className="size-full" src={unsafeUrl} controls />
+          </div>
 
-            <Caption style={{ width: '100%' }} align={align}>
-              <CaptionTextarea
-                className="h-20"
-                readOnly={readOnly}
-                placeholder="Write a caption..."
-              />
-            </Caption>
-          </figure>
-          {children}
-        </PlateElement>
-      );
-    }
-  )
+          <Caption style={{ width: '100%' }} align={align}>
+            <CaptionTextarea
+              className="h-20"
+              readOnly={readOnly}
+              placeholder="Write a caption..."
+            />
+          </Caption>
+        </figure>
+        {children}
+      </PlateElement>
+    );
+  })
 );